Reverted back to "3m" timeframe, plus position_dependency.
Changed the variable, in_position, to reflect actual value in portfolio. Trades on user-provided percentage threshold.

+++ Jupyter/SuperTrend/bnb_bot_v1.py
@@ -56,7 +56,7 @@ in_position = False
 ticker = 'BNB/USD'
 trade_amount = 20
 bar = exchange.fetch_ohlcv(f'{ticker}', timeframe='1m', limit=5)
-order_size = int(trade_amount/bar[4][1]-(.05*(trade_amount/bar[4][1]))) #Trades $1000.
+order_size = float(trade_amount/bar[4][1]-(.05*(trade_amount/bar[4][1]))) #Trades $1000.
 
 #Decision maker.
 def check_buy_sell_signals(df):
